requirements.txt
numpy==1.26.4
scikit-learn>=1.3.0
scikit-surprise>=1.1.3
streamlit>=1.35.0
pandas>=2.0.0
